Basically Done
Ok so it works if no jank stuff happens. But if the system of equations
doesn't have a solution, it does weird stuff. We need to check if the
determinant of the coefficient matrix is zero, and if it is zero, we
need to rerun the program with polynominal degree - 1 (I think).